% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ArdecheRiver}
\alias{ArdecheRiver}
\title{Annual minimum stage record of the Ardèche River at Meyras, France}
\format{
\describe{
\item{Year}{Year}
\item{H}{Stage record}
\item{uH}{Uncertainty expressed as standard deviation (Random realisation)}
}
Details from stage record available in :
}
\source{
\url{https://hydro.eaufrance.fr/stationhydro/V500403001/series}
}
\usage{
ArdecheRiver
}
\description{
A data frame containing annual minimum stages (H, in meters) for the Ardèche River
at Meyras, France, along with by associated random uncertainties from 1988 until 2022.
Assumption: uncertainties follow a normal distribution with a mean of 0.08 meters
and a standard deviation of 0.04 meters.
}
\keyword{datasets}
