% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ArdecheRiverMeyrasGaugings}
\alias{ArdecheRiverMeyrasGaugings}
\title{Gauging of the Ardèche River at Meyras, France, provided by UHPC Grand Delta}
\format{
\describe{
\item{Day}{Day}
\item{Month}{Month}
\item{Year}{Year}
\item{Hour}{Hour}
\item{Minute}{Minute}
\item{Second}{Second}
\item{Date}{Date time}
\item{H}{Stage record}
\item{Q}{Discharge ADCP measurement}
\item{uQ}{Uncertainty expressed as standard deviation}
}
More information about dataset available in :
}
\source{
\url{https://www.theses.fr/2021GRALU006}
}
\usage{
ArdecheRiverMeyrasGaugings
}
\description{
A data frame containing stages (H, in meters) and discharges ADCP measurements
(Q, in cubic meters per second) for the Ardèche River at Meyras, France,
along with the associated uncertainties expressed as standard deviations (uQ) from 2001 until 2018
}
\keyword{datasets}
