% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{BrownianFunk}
\alias{BrownianFunk}
\title{Gombay and Horvath's Brownian function}
\usage{
BrownianFunk(z, d, h, S, alpha)
}
\arguments{
\item{z}{real, critical value}

\item{d}{integer>1, difference of dimension between M0 (no-change) and M1 (single-change)}

\item{h}{real, h variable (see suggested references)}

\item{S}{integer, S variable (noted T in suggested references)}

\item{alpha}{real in (0;1), type-I error level of the test.}
}
\value{
a real number equal to the evaluated Brownian function.
}
\description{
Function to be nullified to compute the critical value when d=1.
See Gombay and Horvath (1994, 1996a, 1996b, 1997), and (Renard)\link{https://hal.science/tel-02588353}
(2006, chapter 3, section I.1.5, p. 101).
}
\keyword{internal}
