% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_public.R
\name{Builder_ControlMatrix}
\alias{Builder_ControlMatrix}
\title{Interactive builder for the hydraulic control matrix}
\usage{
Builder_ControlMatrix(ncontrols)
}
\arguments{
\item{ncontrols}{integer value, number of hydraulic controls}
}
\value{
matrix, hydraulic control. This matrix consists of 0s and 1s
}
\description{
Interactive builder for the hydraulic control matrix
}
