% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_public.R
\name{Builder_Prior_Knowledge}
\alias{Builder_Prior_Knowledge}
\title{Builder function to generate prior information on the parameter as a object}
\usage{
Builder_Prior_Knowledge()
}
\value{
object, prior information on the parameter defined
}
\description{
Builder function to generate prior information on the parameter as a object
}
