% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPrivate.R
\name{character_check}
\alias{character_check}
\title{Check if any column as a NA when convert to numeric format}
\usage{
character_check(donnees)
}
\arguments{
\item{donnees}{data frame}
}
\value{
data frame, logical value. if \code{TRUE}, any NA was detected during transformation.
}
\description{
Check if any column as a NA when convert to numeric format
}
