% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPrivate.R
\name{check_param_distribution}
\alias{check_param_distribution}
\title{Check number of parameters depending of distribution}
\usage{
check_param_distribution(distribution, prior)
}
\arguments{
\item{distribution}{string, distribution}

\item{prior}{real vector, prior information about parameter}
}
\value{
logical, return null if number of parameter does not match with specified distribution
}
\description{
Check number of parameters depending of distribution
}
\keyword{internal}
