% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_private.R
\name{check_recession_modeling}
\alias{check_recession_modeling}
\title{Checks necessary to recession modeling}
\usage{
check_recession_modeling(
  time_rec,
  daterec,
  hrec,
  uHrec,
  indx,
  id_recession_model
)
}
\arguments{
\item{time_rec}{real vector, recession duration relative to the first data detected during the recession}

\item{daterec}{vector, time in POSIXct format mandatory}

\item{hrec}{real vector, stage value of the recessions}

\item{uHrec}{real vector, uncertainty of stage}

\item{indx}{integer, factor used to gather the data of a same recession}

\item{id_recession_model}{integer, position of the model selected in the catalog}
}
\value{
string, error message if conditions not satisfied
}
\description{
Checks necessary to recession modeling
}
\keyword{internal}
