% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPrivate.R
\name{check_square_matrix}
\alias{check_square_matrix}
\title{Check square matrix}
\usage{
check_square_matrix(x)
}
\arguments{
\item{x}{matrix, hydraulic controls}
}
\value{
logical, return null if matrix is not square
}
\description{
Check square matrix
}
