% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CongoRiverMINAN}
\alias{CongoRiverMINAN}
\title{MINAN of the discharge simulation obtained by BaRatinAGE for the Congo River at Brazzaville, Republic of the Congo.}
\format{
\describe{
\item{time}{Date time}
\item{year}{Year}
\item{Q}{MINAN of the Discharge simulation }
\item{uQ}{Uncertainty expressed as standard deviation}
}
}
\usage{
CongoRiverMINAN
}
\description{
A data frame containing discharges simulations (Q, in cubic meters per second)
for the Congo River at Brazzaville, Republic of the Congo, along with
the associated uncertainties expressed as standard deviations (uQ) from 1902 until 2024
}
\keyword{datasets}
