% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPublic.R
\name{control_matrix_builder}
\alias{control_matrix_builder}
\title{Builder hydraulic control matrix}
\usage{
control_matrix_builder(ncontrols)
}
\value{
matrix, hydraulic control
}
\description{
Builder hydraulic control matrix
}
