% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPrivate.R
\name{copy_files_to_folder}
\alias{copy_files_to_folder}
\title{Copy all files from a folder to a another folder}
\usage{
copy_files_to_folder(dir.source, dir.destination)
}
\arguments{
\item{dir.source}{character, source directory}

\item{dir.destination}{character, destination directory}
}
\description{
Copy all files from a folder to a another folder
}
\keyword{internal}
