% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE_fit.R
\name{criticize_rec_model}
\alias{criticize_rec_model}
\title{Criticize recession modeling}
\usage{
criticize_rec_model(
  raw_data,
  data.object,
  residuals,
  summary_MCMC,
  rmse_threshold = 0.3,
  min.weight = 0.1,
  max.weight = 1
)
}
\arguments{
\item{raw_data}{data frame, all stage record}

\item{data.object}{object, model object output by the RBaM function}

\item{residuals}{data frame, residual file obtained using RBaM function}

\item{summary_MCMC}{data frame, MCMC summary file obtained using RBaM function}

\item{rmse_threshold}{real positive value, RMSE threshold to qualify a modeled recession as good}

\item{min.weight}{real positive value, weight given to the first part of the data recession}

\item{max.weight}{real positive value, weight given to the last part of the data recession}
}
\value{
list with the following components :
\enumerate{
\item df.to.segm: data frame, description below:
\itemize{
\item indx: integer value, recession curve index
\item b_estimated: real value, asymptotic height estimation
\item ub_estimated: real value, uncertainty in asymptotic height estimation expressed as a standard deviation
\item date: date, time in POSIXct format. time at which the recession begins
}
\item rec_data.plot.h.dt: data frame, description below:
\itemize{
\item time_rec: real positive value, time recession
\item datarec: date, time in POSIXct format. time of each stage-recession
\item hrec: real value, stage-recession observed
\item uHrec: real value, uncertainty in stage-recession observed expressed as a standard deviation
\item indx: integer value, common index for all stage-recession data of a same recession curve
\item status: string, after using RMSE weighted, recession modeled will be accepted or rejected
}
\item residuals.all.info: data frame, description below:
\itemize{
\item indx: real value, recession curve index
\item X1_obs: real value, duration of recession observed
\item Y1_obs: real value, stage-recession observed
\item Y1_sim: real value, stage-recession simulated
\item uH_obs: real positive value, uncertainty on the stage-recession observed
\item weight: real positive value, weight assigned depending on the recession duration
\item value_rmse: real positive value, RMSE calculation
\item rmse_threshold: real positive value, threshold user-defined to accept or reject some recession simulations
\item status: string, after using RMSE weighted, recession modeled will be accepted or rejected
}
}
}
\description{
Weighted RMSE is used as indicator to qualify the performance of recession modeling.
}
