% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_private.R
\name{custom_format}
\alias{custom_format}
\title{Customized axis label with logarithmic scale}
\usage{
custom_format(x)
}
\arguments{
\item{x}{real values}
}
\value{
character, label in logarithmic scale
}
\description{
Customized axis label with logarithmic scale
}
\keyword{internal}
