% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations_RC_recessions.R
\name{Equation_RC_BaRatinBAC}
\alias{Equation_RC_BaRatinBAC}
\title{Equation of BaRatin model (b-a-c)}
\usage{
Equation_RC_BaRatinBAC(H, a, b, c)
}
\arguments{
\item{H}{real value, stage}

\item{a}{real value, coefficient describing geometry characteristics}

\item{b}{real value, offset}

\item{c}{real value, exponent}
}
\value{
real value, discharge
}
\description{
\deqn{Q(h) = a \cdot (h-b)^{c} \quad \text{for } (h>k) \quad (\text{and } Q=0 \quad \text{if } h \leq b)}
}
