% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations_RC_recessions.R
\name{Equation_RC_Exponential}
\alias{Equation_RC_Exponential}
\title{Equation of exponential regression}
\usage{
Equation_RC_Exponential(H, a, b)
}
\arguments{
\item{H}{real value, stage}

\item{a}{real value, initial state}

\item{b}{real value, growth constant}
}
\value{
real value, discharge
}
\description{
\deqn{Q(h)=a \cdot \exp(b \cdot H)}
}
