% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations_RC_recessions.R
\name{Equation_Rec_BR1}
\alias{Equation_Rec_BR1}
\title{Model 2:}
\usage{
Equation_Rec_BR1()
}
\value{
string, model equation
}
\description{
Recession with one exponential and four parameters: \deqn{\alpha_k \cdot \exp(-\lambda \cdot t^c) + \beta_k}
}
\details{
Parameter's description:
\enumerate{
\item{t real value, time in days}
\item{lambda real value, decay rate constant coefficient 1/(days^c)}
\item{beta_k real value, asymptotic stage in meters}
\item{alpha_k real value, scaling coefficient in meters}
\item{c real value, recession exponent in meters}
}
}
