% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations_RC_recessions.R
\name{EquationRC_LinearRegression}
\alias{EquationRC_LinearRegression}
\title{Equation of Linear interpolation \deqn{Q(h)= a \cdot H + b}}
\usage{
EquationRC_LinearRegression(H, a, b)
}
\arguments{
\item{H}{real value, stage}

\item{a}{real value, slope}

\item{b}{real value, intercept}
}
\value{
real value, discharge
}
\description{
Equation of Linear interpolation \deqn{Q(h)= a \cdot H + b}
}
