% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equations_RC_recessions.R
\name{EquationRec_M3}
\alias{EquationRec_M3}
\title{Model 1:}
\usage{
EquationRec_M3()
}
\value{
string, model equation
}
\description{
Recession model with two exponential and asymptotic: \deqn{\alpha_{1_k} \cdot \exp(-\lambda_1 \cdot t) + \alpha_{2_k} \cdot \exp(-\lambda_2 \cdot t) + \beta_k}
}
\details{
Parameter's description:

\enumerate{
\item{ t real value, time in days}
\item{ alpha_1_k real value, scaling coefficient in meters}
\item{ lambda_1 real value, fast runoff in 1/days}
\item{ alpha_2_k real value, scaling coefficient in meters}
\item{ lambda_2 real value, slow emptying of the aquifer in 1/days}
\item{ beta_k real value, asymptotic stage in meters}
}
}
