% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecessionEstimationModelBaM.R
\name{Estimation_Recession}
\alias{Estimation_Recession}
\title{Catalog of recession models available}
\usage{
Estimation_Recession(
  matched.model,
  temp.folder.Recession = file.path(tempdir(), "BaM", "Recession"),
  ...
)
}
\arguments{
\item{matched.model}{character, model selected by user}

\item{temp.folder.Recession}{directory, temporal folder for recession computation}

\item{...}{optional arguments specific to matched.model defined by user}
}
\value{
list, recession estimation with their uncertainties
}
\description{
Catalog of recession models available
}
