% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_recessions.R
\name{Estimation_Recession_BR1}
\alias{Estimation_Recession_BR1}
\title{Estimate exponential recession model with four parameters}
\usage{
Estimation_Recession_BR1(
  data.object,
  equation_rec_model,
  temp.folder.Recession,
  nCyclesrec = 100,
  burnrec = 0.5,
  nSlimrec = max(nCyclesrec/10, 1)
)
}
\arguments{
\item{data.object}{object, model object output by the RBaM function}

\item{equation_rec_model}{character, equation selected to modelling recessions}

\item{temp.folder.Recession}{directory, temporary directory to write computations}

\item{nCyclesrec}{integer, number of MCMC adaptation cycles. Total number of simulations equal to 100*nCycles}

\item{burnrec}{real between 0 (included) and 1 (excluded), MCMC burning factor}

\item{nSlimrec}{integer, MCMC slim step}
}
\value{
list of following components:
\enumerate{
\item residuals: data frame, description below:
\itemize{
\item X1_obs: real value, duration of recession observed
\item X1_true: real value, unbiased recession duration
\item Y1_obs: real value, stage-recession observed
\item Y1_unbiased: real value, unbiased stage-recession observed
\item Y1_sim: real value, stage-recession simulated
\item Y1_res: real value, residual between observed value and simulated value stage-recession
\item Y1_stdres: real value, residual standar deviation at each stage-recession
}
\item summary_MCMC: data frame, statistical summary of all estimated parameters
}
}
\description{
Estimate exponential recession model with four parameters
}
