% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecessionEstimationModelBaM.R
\name{Estimation_Recession_BR2}
\alias{Estimation_Recession_BR2}
\title{Exponential recession model with four parameters}
\usage{
Estimation_Recession_BR2(
  CalData,
  Ncurves,
  M,
  MCMC,
  t_grid,
  temp.folder,
  remnant_prior
)
}
\arguments{
\item{CalData}{data frame, calibration data used for calibration}

\item{Ncurves}{real positive value, number of recession extracted}

\item{M}{object, model object output by the RBaM function}

\item{MCMC}{data frame, MCMC results}

\item{t_grid}{data frame, recession duration}

\item{temp.folder}{directory, temporary directory to write computations}

\item{remnant_prior}{object, remnant error model output by the RBaM function}
}
\value{
list, recession estimation with their uncertainties
}
\description{
Exponential recession model with four parameters
}
