% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecessionEstimationModelBaM.R
\name{Estimation_Recession_M3}
\alias{Estimation_Recession_M3}
\title{Two exponential recession model with five parameters}
\usage{
Estimation_Recession_M3(
  raw.data,
  data.object,
  current.rec.model,
  temp.folder.Recession,
  nCyclesrec = 100,
  burnrec = 0.5,
  nSlimrec = max(nCyclesrec/10, 1)
)
}
\arguments{
\item{raw.data}{data frame, all stage record}

\item{data.object}{object, model object output by the RBaM function}

\item{current.rec.model}{character, equation selected to modelling recessions}

\item{temp.folder.Recession}{directory, temporary directory to write computations}

\item{nCyclesrec}{integer, number of MCMC adaptation cycles. Total number of simulations equal to 100*nCycles}

\item{burnrec}{real between 0 (included) and 1 (excluded), MCMC burning factor}

\item{nSlimrec}{integer, MCMC slim step}
}
\value{
list, recession estimation with their uncertainties
}
\description{
Two exponential recession model with five parameters
}
