% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatingCurveEquationsModels.R
\name{Exponential_Equation}
\alias{Exponential_Equation}
\title{Equation of exponential regression \deqn{Q(h)=a \cdot \exp(b \cdot H)}}
\usage{
Exponential_Equation(H, a, b)
}
\arguments{
\item{H}{real value, stage}

\item{a}{real value, initial state}

\item{b}{real value, growth constant}
}
\value{
equation, \deqn{Q(h)=a \cdot \exp(b \cdot H)}
}
\description{
Equation of exponential regression \deqn{Q(h)=a \cdot \exp(b \cdot H)}
}
