% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageRecordSegment.R
\name{Extraction_recession}
\alias{Extraction_recession}
\title{Recession extraction}
\usage{
Extraction_recession(
  H,
  uH,
  time,
  filter.H = 1,
  chi = stats::quantile(H, probs = 0.95, na.rm = TRUE),
  delta.t.min = 0,
  delta.t.max = 10,
  tgood = 20,
  Nmin.rec = 10,
  tburn.rec = 0.2
)
}
\arguments{
\item{H}{real vector , stage}

\item{uH}{real vector, uncertainty of the stage}

\item{time}{vector, time in POSIXct format is mandatory}

\item{filter.H}{integer value, filter for number of step records. 10 means only one stage sample every 10 is kept}

\item{chi}{real value, maximum stage rise between two recession data}

\item{delta.t.min}{integer value, minimum days between two recession data}

\item{delta.t.max}{integer value, maximum days between two recession data}

\item{tgood}{integer value, minimum length of the recession in days}

\item{Nmin.rec}{integer value, minimum number of data in a recession curve}

\item{tburn.rec}{numeric, burn factor, >=0 and <1. 0.4 means the first 40 percent of the recession samples are discarded.}
}
\value{
list, all recession satisfying the specified criteria
}
\description{
All recession from the stage record are extracted according to several criteria. NA values are not accepted in the stage record.
}
\examples{

recessions=Extraction_recession(H=ArdecheRiverMeyrasStage$H,
                                uH=0.5,
                                time=ArdecheRiverMeyrasStage$Date,
                                chi=1.5,
                                tgood=30)
plot_rec_extracted(Rec_extracted = recessions)
}
