% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitRC}
\alias{fitRC}
\title{fit}
\usage{
fitRC(time, H, Q, uQ = 0, funk = fitRC_loess, ...)
}
\arguments{
\item{time}{real vector, time}

\item{H}{real vector, stage}

\item{Q}{real vector, discharge}

\item{uQ}{real vector, uncertainty in discharge (as a standard deviation)}

\item{funk}{rating curve model}

\item{...}{input variable of each model}
}
\value{
value
}
\description{
fit
}
