% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitRC_BaRatin}
\alias{fitRC_BaRatin}
\title{Fit rating curve using BaRatin method with prior information multi-hydraulic controls}
\usage{
fitRC_BaRatin(
  time,
  H,
  Q,
  uQ,
  HmaxGrid,
  a.object,
  b.object,
  c.object,
  controlMatrix,
  temp.folder.RC = file.path(tempdir(), "BaM", "RC")
)
}
\arguments{
\item{time}{real vector, time}

\item{H}{real vector, stage}

\item{Q}{real vector, discharge}

\item{uQ}{real vector, uncertainty in discharge (as a standard deviation)}

\item{HmaxGrid}{real value, maximum stage of all data}

\item{a.object}{list of object, created by \code{prior_infor_param_builder} for describing prior information about the geometry properties for each hydraulic control}

\item{b.object}{list of object, created by \code{prior_infor_param_builder} for describing prior information about the offset (thalweg or streambed) for each hydraulic control}

\item{c.object}{list of object, created by \code{prior_infor_param_builder} for describing prior information about the type of hydraulic control for each hydraulic control}

\item{controlMatrix}{matrix, hydraulic control. The function \code{control_matrix_builder} was developed to help the user to created this control matrix}

\item{temp.folder.RC}{directory, temporary directory to write computations of rating curve using observed stages and grid for plotting rating curve}
}
\value{
List with the following components :
\enumerate{
\item ResultsResiduals : data frame, results after fitting curve
\itemize{
\item time: real value, time
\item H: real value, stage
\item Q_obs: real value, discharge observed
\item Q_sim: real value, discharge simulated
\item Q_res: real value, residual between discharge observed and simulated
\item uQ_obs: real value, uncertainty in discharge observed (as a standard deviation)
\item uQ_sim: real value, uncertainty in discharge simulated (as a standard deviation)
}
\item parameters : data frame, parameters of the simplified BaRatin model : \deqn{Q(h)=a \cdot (H - b)^ c}
\itemize{
\item a : list, real value of the parameter of geometry component for each hydraulic control
\item b : list, real value of the parameter of offset (thalweg or streambed) for each hydraulic control
\item c : list, real value of the parameter of exponent describing the type of hydraulic control for each hydraulic control
}
}
}
\description{
Rating curve estimated by the equation \deqn{Q(h)=a \cdot (h-b)^c} with prior information on the parameters. Multi-hydraulic control is considered
}
