% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitRC.R
\name{FitRC_LOESS}
\alias{FitRC_LOESS}
\title{Fit rating curve LOESS model}
\usage{
FitRC_LOESS(time, H, Q, uQ, uH)
}
\arguments{
\item{time}{real vector, time}

\item{H}{real vector, stage}

\item{Q}{real vector, discharge}

\item{uQ}{real vector, uncertainty in discharge (as a standard deviation)}

\item{uH}{real vector, uncertainty in stage record in meters (as a standard deviation)}
}
\value{
list with the following components :
\enumerate{
\item ResultsResiduals : data frame, results after fitting curve
\itemize{
\item time: real value, time
\item H: real value, stage
\item uH: real value, uncertainty in stage observed (as a standard deviation)
\item Q_obs: real value, discharge observed
\item Q_sim: real value, discharge simulated
\item Q_res: real value, residual between discharge observed and simulated
\item uQ_obs: real value, uncertainty in discharge observed (as a standard deviation)
\item uQ_sim: real value, uncertainty in discharge simulated (as a standard deviation)
}
\item parameters : data.frame
\itemize{
\item NULL, because parameters cannot be estimated
}
}
}
\description{
LOESS model used to estimate a simple rating curve
}
\examples{
# Dataset
subset = RhoneRiverGaugings[1:20,]

# LOESS regression to estimate a simple rating curve
fit.funk=FitRC_LOESS(time=subset$Time,H=subset$H,Q=subset$Q,uQ=subset$uQ,uH=0)
fit=fit.funk$ResultsResiduals

# plot rating curve with model results
plot(fit$H, fit$Q_obs, ylim = range(c(fit$Q_obs - fit$uQ_obs, fit$Q_obs + fit$uQ_obs)),
     xlab = "H", ylab = "Q_obs")

# Add error bars following uQ_obs
arrows(fit$H, fit$Q_obs - fit$uQ_obs, fit$H, fit$Q_obs + fit$uQ_obs, angle = 90,
       code = 3, length = 0.1)

# Loess model regression
lines(x=fit$H,y=fit$Q_sim, col="blue",lty=3, lwd=2)

# plot residuals
plot(x=fit$time,y=fit$Q_res, ylim = range(c(fit$Q_res - fit$uQ_obs, fit$Q_res + fit$uQ_obs)),
     xlab='H', ylab='Residual')
arrows(fit$time, fit$Q_res - fit$uQ_obs, fit$time, fit$Q_res + fit$uQ_obs, angle = 90,
       code = 3, length = 0.1)
abline(h=0, col='red')
}
