% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitRecession_M3}
\alias{fitRecession_M3}
\title{Title}
\usage{
fitRecession_M3(
  time_rec,
  hrec,
  uhrec,
  indx,
  nCyclesrec = 100,
  burnrec = 0.5,
  nSlimrec = max(nCycles/10, 1),
  temp.folder.Recession = file.path(tempdir(), "BaM", "Recession")
)
}
\arguments{
\item{time_rec}{real vector, recession duration relative to the first data detected during the recession}

\item{hrec}{real vector, stage value of the recessions}

\item{uhrec}{real vector, uncertainty of stage value of the recessions}

\item{indx}{integer, factor used to gather the data of a same recession}

\item{nCyclesrec}{integer, number of MCMC adaptation cycles. Total number of simulations equal to 100*nCycles}

\item{burnrec}{real between 0 (included) and 1 (excluded), MCMC burning factor}

\item{nSlimrec}{integer, MCMC slim step}

\item{temp.folder.Recession}{directory, temporary directory to write computations}
}
\value{
to complete!!
}
\description{
Title
}
\details{
Prior values are the same for all recession for the parameter a1, a2 and a3
}
\examples{
time_rec = recessions$time_rec
hrec = recessions$hrec
uhrec = recessions$uHrec
indx = recessions$Rec_id
}
