% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_private.R
\name{get_legend}
\alias{get_legend}
\title{Get legend of the ggplot}
\source{
http://www.sthda.com/english/wiki/wiki.php?id_contents=7930#add-a-common-legend-for-multiple-ggplot2-graphs
}
\usage{
get_legend(myggplot)
}
\arguments{
\item{myggplot}{ggplot}
}
\value{
ggplot, legend
}
\description{
Get legend of the ggplot
}
\keyword{internal}
