% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catalog.R
\name{GetCatalog}
\alias{GetCatalog}
\title{RatingShiftHappens catalog}
\usage{
GetCatalog(printOnly = FALSE)
}
\arguments{
\item{printOnly}{Logical, should the catalog be returned or only printed?}
}
\value{
If \code{printOnly==FALSE}, a list with the following fields:
\enumerate{
\item models: available models
}
}
\description{
Rating curve models available
}
\examples{
catalog <- GetCatalog()
GetCatalog(printOnly=TRUE)
}
