% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{getCriticalValue}
\alias{getCriticalValue}
\title{Compute critical value}
\usage{
getCriticalValue(d, alpha, n)
}
\arguments{
\item{d}{integer>1, difference of dimension between M0 (no-change) and M1 (single-change)}

\item{alpha}{real in (0;1), type-I error level of the test.}

\item{n}{integer, sample size}
}
\value{
a real number equal to the critical value of the test.
}
\description{
Compute the critical value of a likelihood-ratio test for a single step change
at an unknown position. Based on the approximations suggested by Gombay and Horvath (1994, 1996a, 1996b, 1997),
as detailled in Renard (2006, chapter 3, section I.1.5, p. 101, https://hal.science/tel-02588353).
}
\keyword{internal}
