% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{getOutputList}
\alias{getOutputList}
\title{Default output list}
\usage{
getOutputList(time, obs, u)
}
\arguments{
\item{time}{real vector, time}

\item{obs}{real vector, observations}

\item{u}{real vector, uncertainty in observations (as a standard deviation)}
}
\value{
a list, see ?Segmentation for details.
}
\description{
Get the default output list returned by Segmentation when failing.
}
\keyword{internal}
