% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPrivate.R
\name{identify_model}
\alias{identify_model}
\title{Identify model selected is in the catalog}
\usage{
identify_model(current_model, model_list)
}
\arguments{
\item{current_model}{name function, model selected}

\item{model_list}{list, models available in the file \code{RecessionCurveEquationsModels.R}}
}
\value{
equation of the model selected
}
\description{
Identify model selected is in the catalog
}
