% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatingCurveEquationsModels.R
\name{LinearInterpolation_Equation}
\alias{LinearInterpolation_Equation}
\title{Equation of Linear interpolation \deqn{Q(h)= a \cdot H + b}}
\usage{
LinearInterpolation_Equation(H, a, b)
}
\arguments{
\item{H}{real value, stage}

\item{a}{real value, slope}

\item{b}{real value, intercept}
}
\value{
equation, \deqn{Q(h)= a \cdot H + b}
}
\description{
Equation of Linear interpolation \deqn{Q(h)= a \cdot H + b}
}
