% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatingCurveEquationsModels.R
\name{LinearRegression_Equation}
\alias{LinearRegression_Equation}
\title{Equation of Linear interpolation \deqn{Q(h)= a \cdot H + b}}
\usage{
LinearRegression_Equation(H, a, b)
}
\arguments{
\item{H}{real value, stage}

\item{a}{real value, slope}

\item{b}{real value, intercept}
}
\value{
real value, discharge
}
\description{
Equation of Linear interpolation \deqn{Q(h)= a \cdot H + b}
}
