% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPrivate.R
\name{List_to_DF}
\alias{List_to_DF}
\title{Convert à list to a data frame if possible}
\usage{
List_to_DF(listx)
}
\arguments{
\item{listx}{list}
}
\value{
data frame or list
}
\description{
Convert à list to a data frame if possible
}
