% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPrivate.R
\name{localmin}
\alias{localmin}
\title{Finding the local minimum}
\usage{
localmin(t, h, uH)
}
\arguments{
\item{t}{vector, time indicating the time of the record}

\item{h}{real vector, stage}

\item{uH}{real vector, uncertainty stage}
}
\value{
data frame, all stage data decreasing
}
\description{
Find all stage data decreasing: h(t) < h(t-1)
}
