% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageRecordSegment.R
\name{model.recession.regression}
\alias{model.recession.regression}
\title{Title}
\usage{
model.recession.regression(
  recessions,
  temp.folder = file.path(tempdir(), "BaM")
)
}
\description{
Title
}
