% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_H_segm_recession}
\alias{plot_H_segm_recession}
\title{Plot segmentation using all stage record}
\usage{
plot_H_segm_recession(time, obs, u, plot_summary)
}
\arguments{
\item{time}{vector, time in POSIXct format is mandatory}

\item{obs}{real vector , stage (m)}

\item{u}{real vector, uncertainty of the stage}

\item{plot_summary}{list, plot data resulting from any segmentation function}
}
\value{
ggplot, plot all stage record after segmentation showing homogeneous periods
}
\description{
Plot segmentation using all stage record
}
\details{
For an example, please see \code{?ModelAndSegmentation.recession.regression}.
}
