% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_modelAndSegm_recession}
\alias{plot_modelAndSegm_recession}
\title{Plot recession modeled and segmentation}
\usage{
plot_modelAndSegm_recession(
  model_rec,
  spec_recession = NULL,
  all_recession = FALSE,
  temp.folder = file.path(tempdir(), "BaM", "Recession"),
  CalibrationData = "CalibrationData.txt",
  fit = fitRecession_M3,
  ...
)
}
\arguments{
\item{model_rec}{list, results obtained by using \code{ModelAndSegmentation.recession.regression} function}

\item{spec_recession}{integer vector, number of recession to plot the observed and simulated recession data separately}

\item{all_recession}{logical, \code{TRUE} = plot all recessions with observed and simulated recession data}

\item{temp.folder}{directory, temporary directory to write computations, be sure to use the same from \code{ModelAndSegmentation.recession.regression} function}

\item{CalibrationData}{character, name of the calibration data used in the \code{ModelAndSegmentation.recession.regression} function. It must to match or an error message will be appear}

\item{fit}{function, fit used during recession modelling}

\item{...}{optional arguments, as vector, to consider shift declared and stored by the hydrometric unit (see \code{?plotSegmentation})}
}
\value{
list of plots
}
\description{
Plot recession modeled and segmentation
}
\details{
If all_recession = \code{TRUE}, spec_recession will not be plotted, because they have already been plotted
}
