% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_Q_ModelAndSegmentation}
\alias{plot_Q_ModelAndSegmentation}
\title{Plot discharge time series after segmentation}
\usage{
plot_Q_ModelAndSegmentation(summary, plot_summary, uH = NA, ...)
}
\arguments{
\item{summary}{list, summary data resulting from model and segmentation function}

\item{plot_summary}{list, plot data resulting from any segmentation function}

\item{uH}{vector or value, uncertainty concerning the observed stage in meters}

\item{...}{optional arguments, as vector, to consider shift declared and stored by the hydrometric unit}
}
\value{
List with the following components :
\enumerate{
\item final_plot: ggplot, observed data presenting the shift time and their estimation
\item observation_and_shift: ggplot, observed data indexed by period, and the estimated shift time is indicated vertically,
\item shift_time_density: ggplot, density plot of the shift times following MCMC results, with
vertical lines indicating the 95\% credibility interval and a red cross representing shift time assignment
}
}
\description{
Plots focusing on the observed data and shift times after using \code{ModelAndSegmentation_Gaugings} function, along with associated uncertainties
}
