% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_segm_recession}
\alias{plot_segm_recession}
\title{Plot recession modeled and segmented}
\usage{
plot_segm_recession(model_rec)
}
\arguments{
\item{model_rec}{list, results obtained by using \code{ModelAndSegmentation.recession.regression} function}
}
\value{
list of plots with the following components:
\enumerate{
\item plot.rec.segmented: ggplot, recession data indexed by period, and  estimated shift time is indicated vertically
\item plot.b.segmented: ggplot, asymptotic height indexed by period, and the estimated shift time is indicated vertically
}
}
\description{
Plot recession modeled and segmented
}
