% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotExtract_Recessions}
\alias{PlotExtract_Recessions}
\title{Plot recession extracted}
\usage{
PlotExtract_Recessions(Rec_extracted, error_bar_plot = FALSE)
}
\arguments{
\item{Rec_extracted}{data frame, recession extracted after using \code{Extract_Recessions}}

\item{error_bar_plot}{logical, \code{TRUE} = plot error bar.}
}
\value{
list, plots of recession extracted. First plot indicates recession following stage record, second plot shows recession time and stage and third plot indicates the minimum recession stage.
}
\description{
Plot recession extracted
}
