% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotGaugingsSegmented}
\alias{plotGaugingsSegmented}
\title{Plot all gaugings after segmentation}
\usage{
plotGaugingsSegmented(
  summary,
  show_gauging = TRUE,
  show_RC = FALSE,
  logscale = FALSE
)
}
\arguments{
\item{summary}{list, summary data resulting from the \code{ModelAndSegmentation_Gaugings} function}

\item{show_gauging}{logical, if \code{TRUE} : gaugings are plotted}

\item{show_RC}{logical, if \code{TRUE} : rating curves estimation are plotted}

\item{logscale}{logical, if \code{TRUE} : log scale is applied}
}
\value{
ggplot, gaugings after segmentation
}
\description{
Plot all gaugings after segmentation
}
