% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotModelAndSegmentation}
\alias{plotModelAndSegmentation}
\title{Plot segmentation with the updated rating curve}
\usage{
plotModelAndSegmentation(summary)
}
\arguments{
\item{summary}{list, summary data resulting from model and segmentation function}
}
\value{
a ggplot
}
\description{
Plot the final segmentation with the updated rating curve,
displaying shift time along with uncertainties
}
\examples{
# Apply recursive segmentation
results=recursive.ModelAndSegmentation(H=RhoneRiver$H,
                                       Q=RhoneRiver$Q,
                                       time=RhoneRiver$Year,
                                       uQ=RhoneRiver$uH)

# plot recursive segmentation
plotModelAndSegmentation(summary=results$summary)
}
