% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotRC_ModelAndSegmentation}
\alias{plotRC_ModelAndSegmentation}
\title{Plot Rating curve after segmentation}
\usage{
plotRC_ModelAndSegmentation(summary, logscale = FALSE)
}
\arguments{
\item{summary}{list, summary data resulting from \code{recursive.ModelAndSegmentation} function}

\item{logscale}{logical, \code{TRUE}= log scale, \code{FALSE} = normal scale}
}
\value{
ggplot, rating curve after segmentation
}
\description{
Plot of all the rating curve after using \code{recursive.ModelAndSegmentation} function, along with associated uncertainties
}
