% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotRCSegmentation_Gaugings}
\alias{PlotRCSegmentation_Gaugings}
\title{Plot Rating curve after segmentation}
\usage{
PlotRCSegmentation_Gaugings(
  summary,
  equation = EquationRC_Exponential,
  ...,
  Hmin_user = 0,
  Hmax_user = 2,
  H_step_discretization = 0.01,
  autoscale = TRUE,
  logscale = FALSE
)
}
\arguments{
\item{summary}{data.frame, summary data resulting from the function \code{ModelAndSegmentation_Gaugings} saved in \verb{$summary}}

\item{equation}{the equation to be considered for plotting : see ‘Details’}

\item{...}{extra arguments for the specified equation : see 'Details'. Respect the name of the arguments (a, b, ..., c, control matrix) depending of the equation used.}

\item{Hmin_user}{real value, minimal stage record set by user for plotting in meters}

\item{Hmax_user}{real value, minimal stage record set by user for plotting in meters}

\item{H_step_discretization}{real positive non-zero value, time step of the discretization in meters}

\item{autoscale}{logical, auto scale following data for plotting}

\item{logscale}{logical, \code{TRUE}= log scale in discharge axis, \code{FALSE} = normal scale in discharge axis}
}
\value{
ggplot, rating curve after segmentation
}
\description{
Plot of all the rating curve after using the function \code{ModelAndSegmentation_Gaugings}, along with associated uncertainties
}
\details{
Some equations for estimating the rating curve are available in this package.
Use \code{GetCatalog()$Equations} to discover the equations supported. More information in \code{?GetCatalog()}.
For the extra information, you can view the arguments through \code{args()}  by enclosing the equation specified in the input arguments in brackets as shown in the example of \verb{?ModelAndSegmentation_Gaugings``. Information about parameters values are available from the function }ModelAndSegmentation_Gaugings\verb{saved in}$summary$param.equation\verb{. Save each parameter estimates separately Please ensure that you enter the same number of rating curve parameters in the extra information as shown in the example }?ModelAndSegmentation_Gaugings`.
}
