% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotresidualModelAndSegmentation}
\alias{plotresidualModelAndSegmentation}
\title{Plot residuals with updating rating curve}
\usage{
plotresidualModelAndSegmentation(summary)
}
\arguments{
\item{summary}{list, summary data resulting from model and segmentation function}
}
\value{
ggplot, residual segmentation
}
\description{
Plot residuals after using \code{recursive.ModelAndSegmentation} function, along with associated uncertainties
}
\examples{
# Apply recursive model and segmentation function
results=recursive.ModelAndSegmentation(H=RhoneRiver$H,
                                       Q=RhoneRiver$Q,
                                       time=RhoneRiver$Year,
                                       uQ=RhoneRiver$uH)

plotresidualModelAndSegmentation(summary=results$summary)
}
