% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotSegmentation_Gaugings_Residual}
\alias{PlotSegmentation_Gaugings_Residual}
\title{Plot residuals with updating rating curve}
\usage{
PlotSegmentation_Gaugings_Residual(summary, plot_summary, ...)
}
\arguments{
\item{summary}{list, summary data resulting from model and segmentation function}

\item{plot_summary}{list, plot data resulting from any segmentation function}

\item{...}{optional arguments, as vector, to consider shift declared and stored by the hydrometric unit}
}
\value{
ggplot, residual segmentation
}
\description{
Plot residuals after using \code{ModelAndSegmentation_Gaugings} function, along with associated uncertainties
}
