% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotSimObs_Recessions}
\alias{PlotSimObs_Recessions}
\title{Comparison between simulated and observed data-recession}
\usage{
PlotSimObs_Recessions(
  model_rec,
  spec_recession = NULL,
  recession_rejected = TRUE,
  all_recession = FALSE
)
}
\arguments{
\item{model_rec}{list, results obtained by using \code{ModelAndSegmentation_Recessions} function}

\item{spec_recession}{integer vector, number of recession to plot the observed and simulated recession data separately}

\item{recession_rejected}{logical, \code{TRUE} = includes plot of all recession rejected}

\item{all_recession}{logical, \code{TRUE} = plot all recessions with observed and simulated recession data}
}
\value{
ggplot, comparison between recession observed and simulated and the status if this is accepted or rejected
}
\description{
Comparison between simulated and observed data-recession
}
\details{
If all_recession = \code{TRUE}, spec_recession will not be plotted, because they have already been plotted
}
