% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{PlotStageSegmentation}
\alias{PlotStageSegmentation}
\title{Plot shift times}
\usage{
PlotStageSegmentation(summary)
}
\arguments{
\item{summary}{list, summary data resulting from model and segmentation function}
}
\value{
ggplot, stage record and shift times
}
\description{
Plot shift times after using \code{recursive.ModelAndSegmentation} function, along with associated uncertainties
}
\examples{
# Apply recursive model and segmentation function
results=recursive.ModelAndSegmentation(H=RhoneRiver$H,
                                       Q=RhoneRiver$Q,
                                       time=RhoneRiver$Year,
                                       uQ=RhoneRiver$uH)

PlotStageSegmentation(summary=results$summary)
}
