% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPublic.R
\name{prior_infor_param_builder}
\alias{prior_infor_param_builder}
\title{Builder function to generate prior information on the parameter as a object}
\usage{
prior_infor_param_builder()
}
\value{
object, prior information on the parameter defined
}
\description{
Builder function to generate prior information on the parameter as a object
}
