% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{quickApprox_getTheta}
\alias{quickApprox_getTheta}
\title{Compute theta given tau}
\usage{
quickApprox_getTheta(tau, time, obs, u, varShift)
}
\arguments{
\item{tau}{real, shift time}

\item{time}{real vector, time}

\item{obs}{real vector, observations}

\item{u}{real vector, uncertainty in observations (as a standard deviation)}

\item{varShift}{logical, allow for a shifting variance?}
}
\value{
A list with the following components:
\enumerate{
\item par: numeric vector, parameter vector: either (mu1,mu2,sigma) (if varShift is FALSE) or
(mu1,mu2,sigma1,sigma2) (if varShift is TRUE)
\item value: real, corresponding log-likelihood
}
}
\description{
Compute theta (mu's and sigma's) given tau (shift time) by maximizing the log-likelihood,
or using explicit estimates if available (which is the case when u==0). Also returns the
associated log-likelihood.
}
\keyword{internal}
