% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{quickApprox_getThetaOnly}
\alias{quickApprox_getThetaOnly}
\title{Compute theta given tau}
\usage{
quickApprox_getThetaOnly(tau, time, obs, u, varShift)
}
\arguments{
\item{tau}{real, shift time}

\item{time}{real vector, time}

\item{obs}{real vector, observations}

\item{u}{real vector, uncertainty in observations (as a standard deviation)}

\item{varShift}{logical, allow for a shifting variance?}
}
\value{
A numeric vector: either (mu1,mu2,sigma) (if varShift is FALSE) or
(mu1,mu2,sigma1,sigma2) (if varShift is TRUE)
}
\description{
Compute theta (mu's and sigma's) given tau (shift time) by maximizing the log-likelihood,
or using explicit estimates if available (which is the case when u==0).
}
\keyword{internal}
