% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{quickApprox_llfunk0}
\alias{quickApprox_llfunk0}
\title{No-shift likelihood}
\usage{
quickApprox_llfunk0(theta, obs, u)
}
\arguments{
\item{theta}{real vector, parameter vector (mu,sigma)}

\item{obs}{real vector, observations}

\item{u}{real vector, uncertainty in observations (as a standard deviation)}
}
\value{
a numeric value equal to the log-likelihood of the no-shift model
}
\description{
No-shift likelihood function for the quick approximation algorithm.
}
\examples{
quickApprox_llfunk0(theta=c(0,1),obs=rnorm(100),u=rep(0,100))
}
\keyword{internal}
