% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{quickApprox_llfunk1}
\alias{quickApprox_llfunk1}
\title{Single-shift likelihood}
\usage{
quickApprox_llfunk1(theta, tau, time, obs, u)
}
\arguments{
\item{theta}{real vector, parameter vector: either (mu1,mu2,sigma) (treated as a fixed-var model) or
(mu1,mu2,sigma1,sigma2) (treated as a shifting-var model)}

\item{tau}{real, shift time}

\item{time}{real vector, time}

\item{obs}{real vector, observations}

\item{u}{real vector, uncertainty in observations (as a standard deviation)}
}
\value{
a numeric value equal to the log-likelihood of the single-shift model
}
\description{
Single-shift likelihood function used in the quick approximation algorithm.
}
\examples{
quickApprox_llfunk1(theta=c(0,1,2),tau=40,time=1:100,obs=c(rnorm(40),rnorm(60)+1),u=rep(0,100))
}
\keyword{internal}
