% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecessionCurveEquationsModels.R
\name{Recession_B2_Equation}
\alias{Recession_B2_Equation}
\title{Model 2:}
\usage{
Recession_B2_Equation()
}
\value{
string, model equation
}
\description{
Recession with one exponential and four parameters: \deqn{\alpha \cdot \exp(-\lambda \cdot t^c) + \beta_k}
}
\details{
Parameter's description:
\enumerate{
\item{t real value, time in days}
\item{lambda_1 real value, decay rate constant coefficient 1/(days^c)}
\item{beta_k real value, asymptotic stage in meters}
\item{alpha_1 real value, scaling coefficient in meters}
\item{c real value, recession exponent in meters}
}
}
