% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecessionCurveEquationsModels.R
\name{Recession_BR1_Equation}
\alias{Recession_BR1_Equation}
\title{Recession with one exponential and five parameters}
\usage{
Recession_BR1_Equation(t, alpha_1, lambda_1, lambda_2, lambda_3, beta_k)
}
\arguments{
\item{t}{real value, time}

\item{alpha_1}{real value, initial stage of the exponential function}

\item{lambda_1}{real value, first parameter of the exponential function}

\item{lambda_2}{real value, second parameter of the exponential function}

\item{lambda_3}{real value, third parameter of the exponential function}

\item{beta_k}{real value, asymptotic stage}
}
\description{
Recession with one exponential and five parameters
}
