% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPrivate.R
\name{Recession_models_available}
\alias{Recession_models_available}
\title{Find recession models}
\usage{
Recession_models_available()
}
\value{
vector, return equation of all recession model available
}
\description{
Find recession models
}
