% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelAndSegment.R
\name{recursive.ModelAndSegmentation}
\alias{recursive.ModelAndSegmentation}
\title{Recursive modelling  and segmentation}
\usage{
recursive.ModelAndSegmentation(
  H,
  Q,
  time = 1:length(H),
  uQ = 0 * Q,
  nSmax = 2,
  nMin = 1,
  nCycles = 100,
  burn = 0.5,
  nSlim = max(nCycles/10, 1),
  temp.folder = file.path(tempdir(), "BaM"),
  funk = fitRC_LinearInterpolation,
  ...
)
}
\arguments{
\item{H}{real vector, stage}

\item{Q}{real vector, discharge}

\item{time}{vector, time in POSIXct, string or numeric format}

\item{uQ}{real vector, uncertainty in discharge (as a standard deviation)}

\item{nSmax}{integer, maximum number of segments to assess}

\item{nMin}{integer, minimum number of observations by segment}

\item{nCycles}{integer, number of MCMC adaptation cycles. Total number of simulations equal to 100*nCycles.}

\item{burn}{real between 0 (included) and 1 (excluded), MCMC burning factor}

\item{nSlim}{integer, MCMC slim step}

\item{temp.folder}{directory, temporary directory to write computations}

\item{funk}{the function for estimating the rating curve to be applied: see ‘Details’}

\item{...}{optional arguments to funk}
}
\value{
List with the following components :
\enumerate{
\item summary: list, summarize the information to present to the user
\itemize{
\item data: data frame, all data of (H,Q and uQ) with their respective periods after segmentation
\item shift: data frame, all detected shift time
\item param.equation: data frame, parameters estimation
}
\item res: list, provide all the information of the periods from tree structure
\itemize{
\item tau: real vector, estimated shift times
\item segments: list, segment maximum a posterior (MAP) value indexed by the list number
\item mcmc: data frame, MCMC simulation
\item data.p: list, separate and assign information by identified stable period indexed by the list number
\item DIC: real, DIC estimation
\item nS: integer, optimal number of segments following DIC criterion
}
\item tree: data frame, provide tree structure
}
}
\description{
Recursive procedure for an \strong{unknown} number of segments, updating the rating curve at each iteration before segmentation
}
\details{
Some functions for estimating the rating curve are available in this package. Use \code{?GetCatalog()} to discover the supported functions
}
\examples{
# Apply recursive segmentation
results=recursive.ModelAndSegmentation(H=ArdecheRiverMeyrasGaugings$H,Q=ArdecheRiverMeyrasGaugings$Q,
                                       time=ArdecheRiverMeyrasGaugings$Date,
                                       uQ=ArdecheRiverMeyrasGaugings$uQ,
                                       nSmax=2,nMin=2,funk=fitRC_exponential)

# Data information
knitr::kable(head(results$summary$data),
             align = 'c',row.names = FALSE)

# Shift information
knitr::kable(head(results$summary$shift),
             align = 'c',row.names = FALSE)

# Parameters estimation of the rating curve
results$summary$param.equation

# Have a look at recursion tree
results$tree

# Visualize tree structure
plotTree(results$tree)
# Plot segmentation
plotRC_ModelAndSegmentation(summary=results$summary)
plotRC_ModelAndSegmentation(summary=results$summary,logscale=TRUE)
plotStage_ModelAndSegmentation(summary=results$summary)
plotResidual_ModelAndSegmentation(summary=results$summary)
}
