% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{Recursive_Segmentation}
\alias{Recursive_Segmentation}
\title{Recursive segmentation}
\usage{
Recursive_Segmentation(
  obs,
  time = 1:length(obs),
  u = 0 * obs,
  nSmax = 2,
  nMin = 1,
  nCycles = 100,
  burn = 0.5,
  nSlim = max(nCycles/10, 1),
  temp.folder = file.path(tempdir(), "BaM"),
  mu_prior = list(),
  ...
)
}
\arguments{
\item{obs}{real vector, observations}

\item{time}{real vector, time in POSIXct, string or numeric format}

\item{u}{real vector, uncertainty in observations (as a standard deviation)}

\item{nSmax}{integer, maximum number of segments to assess}

\item{nMin}{integer, minimum number of observations by segment}

\item{nCycles}{integer, number of MCMC adaptation cycles. Total number of simulations equal to 100*nCycles.}

\item{burn}{real between 0 (included) and 1 (excluded), MCMC burning factor}

\item{nSlim}{integer, MCMC slim step}

\item{temp.folder}{directory, temporary directory to write computations}

\item{mu_prior}{list, object describing prior knowledge about residual between the rating curve and observation if user-defined (see details)}

\item{...}{other arguments passed to RBaM::BaM.}
}
\value{
list with the following components:
\enumerate{
\item summary: list, summarize the information to present to the user
\itemize{
\item data: data frame, all data of (H,Q and uQ) with their respective periods after segmentation
\item shift: data frame, all detected shift time
}
\item plot : list, data formatted to use as input for some plot functions
\item res: list, provide all the information of the periods from tree structure
\itemize{
\item tau: real vector, estimated shift times
\item segments: list, segment maximum a posterior (MAP) value indexed by the list number
\item mcmc: data frame, MCMC simulation
\item data.p: list, separate and assign information by identified stable period indexed by the list number
\item DIC: real, DIC estimation
\item nS: integer, optimal number of segments following DIC criterion
}
\item tree: data frame, provide tree structure
\item origin.date: positive real or date, date describing origin of the segmentation for a sample. Useful for recursive segmentation.
}
}
\description{
Recursive segmentation procedure for an \strong{unknown} number of segments
}
\details{
User may enter prior knowledge about the mu parameter (see \code{?Segmentation_Engine}) instead of default values.
This information must be provided using the parameter function in the RBaM package, as shown in the example.
}
\examples{
# Apply recursive segmentation
results=Recursive_Segmentation(obs=RhoneRiverAMAX$H,
                               time=RhoneRiverAMAX$Year,
                               u=RhoneRiverAMAX$uH,
                               nSmax=3)

# Data information
knitr::kable(head(results$summary$data),
             align = 'c',row.names = FALSE)
# Shift information
knitr::kable(head(results$summary$shift),
             align = 'c',row.names = FALSE)
# Have a look at recursion tree
results$tree

# Visualize tree structure
PlotTree(results$tree)
# Plot segmentation
PlotSegmentation(summary=results$summary,
                 plot_summary=results$plot)
}
