% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPrivate.R
\name{remove_files}
\alias{remove_files}
\title{Remove files excepting some specific files from the directory}
\usage{
remove_files(dir.source, files_to_keep = NULL)
}
\arguments{
\item{dir.source}{character, source directory}

\item{files_to_keep}{character, files to keep}
}
\description{
Remove files excepting some specific files from the directory
}
\keyword{internal}
