% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_private.R
\name{replace_negatives_or_zero_values}
\alias{replace_negatives_or_zero_values}
\title{Replace negative and/or zero values to NA or a specified value}
\usage{
replace_negatives_or_zero_values(
  data_frame,
  columns = "all",
  consider_zero = TRUE,
  replace = NA
)
}
\arguments{
\item{data_frame}{data frame, data to be checked and replaced}

\item{columns}{string, exact names of the columns from data frame}

\item{consider_zero}{logical, if \code{TRUE} zero values will be replace by NA. Otherwise, zero values are not be replaced.}

\item{replace}{NA or real value, data will be replace when condition are not accepted}
}
\value{
data frame transformed with NA values
}
\description{
Replace negative and/or zero values to NA or a specified value
}
\keyword{internal}
