% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RhoneRiverAMAX}
\alias{RhoneRiverAMAX}
\title{Floods of the Rhone River at Beaucaire, France}
\format{
\describe{
\item{Year}{Year}
\item{Date}{Date}
\item{Time}{Time}
\item{H}{Stage record}
\item{uH}{Uncertainty of the stage expressed as standard deviation}
\item{Q}{Discharge ADCP measurement}
\item{uQ}{Uncertainty of the discharge expressed as standard deviation}
}
}
\source{
\url{https://doi.org/10.1016/j.jhydrol.2023.129840}
}
\usage{
RhoneRiverAMAX
}
\description{
A data frame containing annual maximum stages (H, in meters) and discharges
(Q, in cubic meters per second) for the Rhone River at Beaucaire, France,
along with the associated uncertainties expressed as standard deviations (uH and uQ).
Details on the reconstruction of of these long series can be found in the article
by Lucas et al. (2023) referenced below.
Note that years 1968, 1969 and 1970 are missing and are not included in the data frame.
}
\keyword{datasets}
