% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RhoneRiverStage}
\alias{RhoneRiverStage}
\title{Daily stage record of the Rhone River at Beaucaire, France}
\format{
An object of class \code{data.frame} with 29772 rows and 2 columns.
}
\usage{
RhoneRiverStage
}
\description{
A data frame containing daily stages (H, in meters) and date
for the Rhone River at Beaucaire, France. Any uncertainty has been reported.
}
\keyword{datasets}
