% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_public.R
\name{RMSE_Weighted}
\alias{RMSE_Weighted}
\title{Root Mean Squared Error weighted (RMSE weighted)}
\usage{
RMSE_Weighted(observations, predicted, weights)
}
\arguments{
\item{observations}{vector, observed data}

\item{predicted}{vector, predicted data}

\item{weights}{vector, weights to each data to give more or less weights to some data}
}
\value{
real value, RMSE weighted value to qualify the fit
}
\description{
Root Mean Squared Error weighted (RMSE weighted)
}
