% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{Segmentation_quickApprox}
\alias{Segmentation_quickApprox}
\title{Segmentation engine - quick approximation algorithm}
\usage{
Segmentation_quickApprox(
  obs,
  time = 1:length(obs),
  u = 0 * obs,
  varShift = FALSE,
  nS = 2,
  nMin = ifelse(varShift, 2, 1),
  nSim = 500
)
}
\arguments{
\item{obs}{real vector, observations}

\item{time}{vector, time in POSIXct, string or numeric format}

\item{u}{real vector, uncertainty in observations (as a standard deviation)}

\item{varShift}{logical, allow for a shifting variance?}

\item{nS}{integer, number of segments, either 1 or 2}

\item{nMin}{integer, minimum number of observations by segment}

\item{nSim}{integer, number of simulated tau values}
}
\value{
list with the following components:
\enumerate{
\item summary: list, summarize the information to present to the user
\itemize{
\item data: data frame, data augmented with a column denoting the period after segmentation
\item shift: data frame, detected shift time in numeric or POSIXct format in UTC
}
\item plot : list, data formatted to use as input for some plot functions
\itemize{
\item density.tau: data frame, a table with three columns. The first column indicates the specific shift being analyzed.
The second column contains the value of the shift time tau. The last columns shows the probability density associated
with each tau value
\item density.inc.tau: data frame, all information about the 95\% credibility interval and the Maximum a posterior (MAP) estimation
for shift times with their associated probability densities
}
\item tau: real, estimated shift time in numeric or POSIXct format in UTC
\item segments: list, segment mean value indexed by the list number
\item mcmc: data frame, MCMC simulation. Note that values for mu's and sigma's are fixed to the values corresponding
to the maxpost tau.
\item data.p: list, separate and assign information by identified stable period indexed by the list number
\item DIC: real, DIC estimation
\item origin.date.p: positive real or date, date describing origin of the segmentation for a sample. Useful for recursive segmentation.
}
}
\description{
A quick approximation to the segmentation procedure for either one or two segments, and
no prior information. The approximation is based on ignoring the uncertainty in mu1, mu2 and sigma,
and focusing on the uncertainty in tau. To achieve this, point-estimates of mu1, mu2 and sigma are
plugged in for each possible value of tau. The resulting algorithm is MCMC-free and independent of RBaM.
}
\examples{
# Run segmentation engine function at two segments
# for data set : RhoneRiverAMAX (further details on ?RhoneRiverAMAX)
res=Segmentation_quickApprox(obs=RhoneRiverAMAX$H,time=RhoneRiverAMAX$Year,
                             u=RhoneRiverAMAX$uH,nS=2)
res$summary$shift
PlotSegmentation(res$summary,res$plot)
}
