% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_public.R
\name{TransformDateFormat}
\alias{TransformDateFormat}
\title{Date to numeric format}
\usage{
TransformDateFormat(date)
}
\arguments{
\item{date}{vector, date in POSIXct, character or Date format}
}
\value{
List with the following components :
\enumerate{
\item origin: value, denoting the oldest date from the input variable
\item time: real vector, representing time in numeric format. This denotes the number of days from origin
}
}
\description{
Get date format and transform date to numeric
}
\examples{
TransformDateFormat(
                   c(as.POSIXct('2024-04-19 12:30:00',
                     tz='UTC'),
                     as.POSIXct('2024-03-19 18:30:00',
                     tz='UTC')))

TransformDateFormat(c(as.Date('2024/02/19'),
                      as.Date('2024/02/10')))

TransformDateFormat(c(as.character('20240419'),
                      as.character('20240119')))
}
