% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{transformDatesInOutput}
\alias{transformDatesInOutput}
\title{Date transformer}
\usage{
transformDatesInOutput(out, origin.date)
}
\arguments{
\item{out}{list, output from Segmentation (see ?Segmentation for details)}

\item{origin.date}{date, origin used to transform date back}
}
\value{
a list, see ?Segmentation for details. It is the same as the list \code{out},
but with all dates reformatted.
}
\description{
Transform the dates contained in a raw output list back into their original format.
}
\keyword{internal}
