% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WairauRiverGaugings}
\alias{WairauRiverGaugings}
\title{Gauging of the Wairau River at Barnetts Bank, Tuamarina, New Zealand, Tideda software export.}
\format{
\describe{
\item{Date}{Date time}
\item{H}{Stage record}
\item{Q}{Discharge ADCP measurement}
\item{uQ}{Uncertainty expressed as standard deviation}
}
}
\usage{
WairauRiverGaugings
}
\description{
A data frame containing stages (H, in meters) and discharges ADCP measurements
(Q, in cubic meters per second) for the Wairau River at Barnetts Bank, Tuamarina, New Zealand,
from 1999 until 2015.
Any information about uncertainty in discharge measurement; hence a uncertainty of 8\%
has been considered and it was expressed as standard deviations (uQ).
}
\keyword{datasets}
