% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checks.R
\name{check_vector_lengths}
\alias{check_vector_lengths}
\title{Check size vectors}
\usage{
check_vector_lengths(...)
}
\arguments{
\item{...}{real vectors}
}
\value{
logical, return null if the vectors have not the same length
}
\description{
Function to check if the vectors have the same length
}
\examples{

check_vector_lengths(RhoneRiver$Time,RhoneRiver$H,RhoneRiver$Q,RhoneRiver$uQ)
check_vector_lengths(RhoneRiver$Time,RhoneRiver$H,RhoneRiver$Q,RhoneRiver$uQ[-1])
}
