% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPrivate.R
\name{check_vector_lengths}
\alias{check_vector_lengths}
\title{Check size vectors}
\source{
\url{https://www.simonqueenborough.info/R/basic/lessons/lapply_and_sapply.html}
}
\usage{
check_vector_lengths(...)
}
\arguments{
\item{...}{real vectors}
}
\value{
logical, return null if the vectors have not the same length
}
\description{
Function to check if the vectors have the same length
}
