% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPublic.R
\name{control_matrix_builder}
\alias{control_matrix_builder}
\title{Interactive builder for the hydraulic control matrix}
\usage{
control_matrix_builder(ncontrols)
}
\arguments{
\item{ncontrols}{integer value, number of hydraulic controls}
}
\value{
matrix, hydraulic control. This matrix consists of 0s and 1s
}
\description{
Interactive builder for the hydraulic control matrix
}
