% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_private.R
\name{convert_list_to_dataframe}
\alias{convert_list_to_dataframe}
\title{Convert list to data frame}
\usage{
convert_list_to_dataframe(liste_df)
}
\arguments{
\item{liste_df}{list, data frame}
}
\value{
data frame if possible, if not a list
}
\description{
Convert list to data frame
}
\keyword{internal}
