% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{DateFormatTransform}
\alias{DateFormatTransform}
\title{Date to time format}
\usage{
DateFormatTransform(date)
}
\arguments{
\item{date}{vector, date}
}
\value{
List with the following components :
\enumerate{
\item date: vector, date in POSIXct format
\item time: real vector, time in numeric format where origin default is "1970-01-01 00:00:00 UTC"
}
}
\description{
Get date format and transform date to numeric
}
\examples{
datetime=DateFormatTransform('2024-02-19 12:30:00')
datetime
as.POSIXct(datetime, origin = '1970-01-01',tz='UTC')

datetime2=DateFormatTransform('2024/02/19')
datetime2
as.POSIXct(datetime2, origin = '1970-01-01',tz='UTC')
}
