% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitRC_BaRatin}
\alias{fitRC_BaRatin}
\title{BaRatin model}
\usage{
fitRC_BaRatin(time, H, Q, uQ, matrix, priors, remnant)
}
\arguments{
\item{time}{real vector, time}

\item{H}{real vector, stage}

\item{Q}{real vector, discharge}

\item{uQ}{real vector, uncertainty in discharge (as a standard deviation)}

\item{matrix}{real data frame, hydraulic controls specified as bonnifait matrix}

\item{priors}{real list, prior information of each parameter}

\item{remnant}{real vector, remnant model information}
}
\value{
real data frame
}
\description{
Rating curve estimated by using BaRatin approach
}
