% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitRC_SimplifiedBaRatin}
\alias{fitRC_SimplifiedBaRatin}
\title{Fit rating curve using simplified BaRatin method (b-a-c) with non-informative prior information with only a rectangular hydraulic control}
\usage{
fitRC_SimplifiedBaRatin(
  time,
  H,
  Q,
  uQ,
  uH,
  HmaxGrid,
  nCycles_RC = 100,
  burn_RC = 0.5,
  nSlim_RC = max(nCycles_RC/10, 1),
  temp.folder.RC = file.path(tempdir(), "BaM", "RC")
)
}
\arguments{
\item{time}{real vector, time}

\item{H}{real vector, stage}

\item{Q}{real vector, discharge}

\item{uQ}{real vector, uncertainty in discharge (as a standard deviation)}

\item{uH}{real vector, uncertainty in stage record in meters (as a standard deviation)}

\item{HmaxGrid}{real value, maximum stage of all data in the historical record}

\item{nCycles_RC}{integer, number of MCMC adaptation cycles. Total number of simulations equal to 100*nCycles}

\item{burn_RC}{real between 0 (included) and 1 (excluded), MCMC burning factor}

\item{nSlim_RC}{integer, MCMC slim step}

\item{temp.folder.RC}{directory, temporary directory to write computations of rating curve using observed stages and grid for plotting rating curve}
}
\value{
List with the following components :
\enumerate{
\item ResultsResiduals : data frame, results after fitting curve
\itemize{
\item time: real value, time
\item H: real value, stage
\item uH: real value, uncertainty in stage observed (as a standard deviation)
\item Q_obs: real value, discharge observed
\item Q_sim: real value, discharge simulated
\item Q_res: real value, residual between discharge observed and simulated
\item uQ_obs: real value, uncertainty in discharge observed (as a standard deviation)
\item uQ_sim: real value, uncertainty in discharge simulated (as a standard deviation)
}
\item parameters : data frame, parameters of the simplified BaRatin model : \deqn{Q(h)=a \cdot (H - b)^ c}
\itemize{
\item a : real value, parameter of the geometry component
\item b : real value, parameter of the offset (thalweg or streambed)
\item c : real value, parameter of the exponent describing the type of hydraulic control
\item k : real value, parameter of the activation stage for each hydraulic control
\item gamma1 ; real value, parameter describing the standard deviation of structural errors (relative to discharge)
\item gamma2 : real value, parameter describing the standard deviation of structural errors (constant error)
}
}
}
\description{
Rating curve estimated by the equation \deqn{Q(h)=a \cdot (h-b)^c} non-informative prior information on the parameters estimated automatically, using stage record and gauging information. Only a rectangular hydraulic control is considered
}
