% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fitRecession_M3}
\alias{fitRecession_M3}
\title{Recession model with two exponential and asymptotic}
\usage{
fitRecession_M3(
  time_rec,
  hrec,
  uHrec,
  indx,
  alpha1.object = NULL,
  alpha2.object = NULL,
  beta.object = NULL,
  lambda1.object = NULL,
  lambda2.object = NULL,
  Dataset.object = NULL,
  nCyclesrec = 100,
  burnrec = 0.5,
  nSlimrec = max(nCyclesrec/10, 1),
  temp.folder.Recession = file.path(tempdir(), "BaM", "Recession")
)
}
\arguments{
\item{time_rec}{real vector, recession duration relative to the first data detected during the recession}

\item{hrec}{real vector, stage value of the recessions}

\item{uHrec}{real vector, uncertainty of stage value of the recessions}

\item{indx}{integer, factor used to gather the data of a same recession}

\item{alpha1.object}{object, prior knowledge on the parameter representing the initial stage of the first exponential function}

\item{alpha2.object}{object, prior knowledge on the parameter representing the initial stage of the second exponential function}

\item{beta.object}{object, prior knowledge on the parameter representing the asymptotic stage}

\item{lambda1.object}{object, prior knowledge on the  parameter describing the fast runoff}

\item{lambda2.object}{object, prior knowledge on the parameter describing the slow emptying of the aquifer}

\item{Dataset.object}{object, dataset given by user}

\item{nCyclesrec}{integer, number of MCMC adaptation cycles. Total number of simulations equal to 100*nCycles}

\item{burnrec}{real between 0 (included) and 1 (excluded), MCMC burning factor}

\item{nSlimrec}{integer, MCMC slim step}

\item{temp.folder.Recession}{directory, temporary directory to write computations}
}
\value{
List with the following components :
\enumerate{
\item ResultsResiduals : list, results after fitting the recession model to each recession observed
\itemize{
\item time: real value, time
\item H: real value, stage observed
\item H_rec_sim: real value, stage simulated
\item H_res: real value, residual between stage observed and simulated
\item uH_rec: real value, uncertainty in stage observed (as a standard deviation)
\item uH_sim: real value, uncertainty in stage simulated (as a standard deviation)
}
\item parameters : data frame, parameters of the recession model
\itemize{
\item alpha1 : real value, parameter representing the initial stage
\item lambda1 : real value, parameter describing the fast runoff
\item alpha2 : real value, parameter representing the initial stage
\item lambda2 : real value, parameter describing the slow emptying of the aquifer
\item beta : real value, parameter representing the asymptotic stage
\item gamma1 ; real value, parameter describing the standard deviation of structural errors (relative to discharge)
\item gamma2 : real value, parameter describing the standard deviation of structural errors (constant error)
}
}
}
\description{
Recession modelling following a the exponential function specified as M3 according to (Darienzo, 2022):
\deqn{h(t) = \alpha_1(k) \cdot \exp(-\lambda_1 \cdot t) + \alpha_2(k) \cdot \exp(-\lambda_2 \cdot t) + \beta}
This model includes three recession-specific parameters: \eqn{\alpha_1}, \eqn{\alpha_2}, and \eqn{\beta}
and two stable parameters: \eqn{\lambda_1} and \eqn{\lambda_2}.
}
\details{
Default values for this recession are shown in ‘Details’

By default, prior values are the same for all recession for the parameter alpha1, alpha2 and beta

Default values are:
\describe{
\item{Starting in 200, we assume that \eqn{\alpha_1 \sim U(0, 1000).}}
\item{Starting in 50, we assume that \eqn{\alpha_2 \sim U(0, 500).}}
\item{Starting in \eqn{\exp(-\log(0.5) + \log(\log(2)))}, we assume that
\eqn{\lambda_1 \sim \text{LogNormal}(\mu = -\log(0.5) + \log(\log(2)), \sigma = 1).}}
\item{Starting in \eqn{\exp(-\log(50) + \log(\log(2)))}, we assume that
\eqn{\lambda_2 \sim \text{LogNormal}(\mu = -\log(80) + \log(\log(2)), \sigma = 0.5).}}
\item{Starting in 1000, we assume that \eqn{\beta \sim U(-10000, 10000).}}
}
}
