% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catalog.R
\name{GetCatalog_RC}
\alias{GetCatalog_RC}
\title{Catalog of rating curves}
\usage{
GetCatalog_RC(printOnly = FALSE)
}
\arguments{
\item{printOnly}{Logical, should the catalog be returned or only printed?}
}
\value{
If \code{printOnly==FALSE}, a list with the following fields:
\enumerate{
\item models: available models
\item equations: available RC equations
}
}
\description{
Rating curve and models with their equation available
}
\examples{
catalog <- GetCatalog_RC()
GetCatalog_RC(printOnly=TRUE)
}
