% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Catalog.R
\name{GetCatalog_Recession}
\alias{GetCatalog_Recession}
\title{RatingShiftHappens catalog of recessions model}
\usage{
GetCatalog_Recession()
}
\value{
list with the following fields depending on the model chosen:
\enumerate{
\item Equation: recession equation
\item funk: recession estimation
}
}
\description{
Available recession equations and their estimation
}
\examples{
catalog <- GetCatalog_Recession()
# Example for the first model in the catalog
catalog[[1]]$Equation
catalog[[1]]$funk
}
