% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageRecordSegment.R
\name{ModelAndSegmentation.recession.regression}
\alias{ModelAndSegmentation.recession.regression}
\title{Model and segmentation of recession}
\usage{
ModelAndSegmentation.recession.regression(
  time_rec,
  hrec,
  uhrec,
  indx,
  nSmax = 2,
  nMin = 1,
  nCycles = 100,
  burn = 0.5,
  nSlim = max(nCycles/10, 1),
  temp.folder = file.path(tempdir(), "BaM"),
  funk = fitRecession_M3,
  ...
)
}
\arguments{
\item{time_rec}{real vector, recession duration relative to the first data detected during the recession}

\item{hrec}{real vector, stage value of the recessions}

\item{uhrec}{real vector, uncertainty of stage value of the recessions}

\item{indx}{integer, factor used to gather the data of a same recession}

\item{nSmax}{integer, maximum number of segments to assess}

\item{nMin}{integer, minimum number of observations by segment}

\item{nCycles}{integer, number of MCMC adaptation cycles. Total number of simulations equal to 100*nCycles}

\item{burn}{real between 0 (included) and 1 (excluded), MCMC burning factor}

\item{nSlim}{integer, MCMC slim step}

\item{temp.folder}{directory, temporary directory to write computations}

\item{funk}{the function for estimating the recession}

\item{...}{optional arguments to funk}
}
\value{
to complete!
}
\description{
Modelling recession using a catalog of fit models available to apply segmentation procedure.
To get the catalog of models available using \code{GetCatalog()}, all functions specifying Recession in the fit are concerned.
For more information about recession model and their default values, please go to the fit function, e.g. \code{?fitRecession_M3}
}
\examples{
recessions=Extraction_recession(H=ArdecheRiverMeyrasStage$H,
                                uH=0.5,
                                time=ArdecheRiverMeyrasStage$Date,
                                chi=1.5,
                                tgood=30)

D=RBaM::dataset(X=recessions['time_rec'],
                Y=recessions['hrec'],
                Yu=recessions['uHrec'],
                VAR.indx=recessions['indx'],
                data.dir=file.path(tempdir(),'BaM','Recession'))

# Give prior knowledge about recession-specific parameters and keep prior on stable parameters
alpha1.object = RBaM::parameter_VAR(name='alpha1',
                                    index='indx',
                                    d=D,
                                    init=rep(100,max(recessions$indx)),
                                    prior.dist=rep('Uniform',max(recessions$indx)),
                                    prior.par=rep(list(c(0,1000)),max(recessions$indx)))

alpha2.object = RBaM::parameter_VAR(name='alpha2',
                                    index='indx',
                                    d=D,
                                    init=rep(50,max(recessions$indx)),
                                    prior.dist=rep('Uniform',max(recessions$indx)),
                                    prior.par=rep(list(c(0,100)),max(recessions$indx)))

beta.object = RBaM::parameter_VAR(name='beta',
                                    index='indx',
                                    d=D,
                                    init=rep(52.4,max(recessions$indx)),
                                    prior.dist=rep('Uniform',max(recessions$indx)),
                                    prior.par=rep(list(c(-97.6,102.4)),max(recessions$indx)))

model_rec=ModelAndSegmentation.recession.regression(time_rec=recessions$time_rec,
                                                    hrec=recessions$hrec,
                                                    uhrec=recessions$uHrec,
                                                    indx=recessions$indx,
                                                    alpha1.object=alpha1.object,
                                                    alpha2.object=alpha2.object,
                                                    beta.object=beta.object,
                                                    Dataset.object=D,
                                                    nCyclesrec=100,
                                                    burnrec=0.5,
                                                    nSlim=10)

}
