% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageRecordSegment.R
\name{ModelAndSegmentation.recession.regression}
\alias{ModelAndSegmentation.recession.regression}
\title{Model and segmentation of recession}
\usage{
ModelAndSegmentation.recession.regression(
  time_rec,
  daterec,
  hrec,
  uHrec,
  indx,
  nSmax = 3,
  nMin = 1,
  nCycles = 100,
  burn = 0.5,
  nSlim = max(nCycles/10, 1),
  temp.folder = file.path(tempdir(), "BaM"),
  temp.folder.Recession = file.path(tempdir(), "BaM", "Recession"),
  funk = Recession_B2_Equation,
  ...
)
}
\arguments{
\item{time_rec}{real vector, recession duration relative to the first data detected during the recession}

\item{daterec}{vector, time in POSIXct format mandatory}

\item{hrec}{real vector, stage value of the recessions}

\item{uHrec}{real vector, uncertainty of stage}

\item{indx}{integer, factor used to gather the data of a same recession}

\item{nSmax}{integer, maximum number of segments to assess}

\item{nMin}{integer, minimum number of observations by segment}

\item{nCycles}{integer, number of MCMC adaptation cycles. Total number of simulations equal to 100*nCycles}

\item{burn}{real between 0 (included) and 1 (excluded), MCMC burning factor}

\item{nSlim}{integer, MCMC slim step}

\item{temp.folder}{directory, temporary directory to write computations of the segmentation}

\item{temp.folder.Recession}{directory, temporary directory to write computations of the recession}

\item{funk}{the equation for estimating the recession}

\item{...}{optional arguments to funk}
}
\value{
List with the following components:
\enumerate{
\item summary.rec.extracted : list with this sub components:
\itemize{
\item data: data frame, all data of recession (time, date, H and uH) with their respective period after segmentation and status if the recession model was accepted or rejected
\item shift: data frame, all detected shift time
}
\item summary.results.segm: list with this sub components:
\itemize{
\item data: data frame, asymptotic height estimation by curve accepted with their respective period after segmentation
\item shift: data frame, all detected shift time
}
\item summary.residual: data frame, the residual information between the MAP simulation and the observation with weighted RMSE to criticise the recession estimate and finally the acceptance or rejection status.
\item plots: list, data formatted to use as input for some plot functions
\item res: list, provide all the information of the periods from tree structure
\itemize{
\item tau: real vector, estimated shift times
\item segments: list, segment maximum a posterior (MAP) value indexed by the list number
\item mcmc: data frame, MCMC simulation
\item data.p: list, separate and assign information by identified stable period indexed by the list number
\item DIC: real, DIC estimation
\item nS: integer, optimal number of segments following DIC criterion
}
\item tree: data frame, provide tree structure
\item origin.date: positive real or date, date describing origin of the segmentation for a sample
}
}
\description{
Modelling recession using a catalog of fit models available to apply segmentation procedure.
Please see details for understading how to use this function
}
\details{
To get the catalog of models available using \code{GetCatalog()$Recession_Equations}.
Be careful when specifying temporal folder for computations.
temp.folder must contain temp.folder.Recession as the default value:
temp.folder= \code{file.path(tempdir(),'BaM')}
temp.folder.Recession= \code{file.path(tempdir(),'BaM','Recession')}
}
\examples{
recessions_extracted=Extraction_recession(H=ArdecheRiverMeyrasStage$H,
                                          uH=0.05,
                                          time=ArdecheRiverMeyrasStage$Date,
                                          chi=0.5,
                                          tgood=30)

recessions = recessions_extracted$Rec_extracted

# Plot all recession extracted
plot_rec_extracted(Rec_extracted = recessions)

# Choose fit recession model
fit=Recession_B2_Equation

model_rec=ModelAndSegmentation.recession.regression(time_rec=recessions$time_rec,
                                                    daterec=recessions$date,
                                                    hrec=recessions$hrec,
                                                    uHrec=recessions$uHrec,
                                                    indx=recessions$indx,
                                                    funk=fit)
# Plot recession segmentation
plot_segm_recession(model_rec=model_rec,
                    spec_recession=c(2,16,28,48),
                    recession_rejected=FALSE,
                    )

# Plot using all data of stage record
plot_H_segm_recession(time=ArdecheRiverMeyrasStage$Date,
                      obs=ArdecheRiverMeyrasStage$H,
                      u=0.05,
                      plot_summary = model_rec$plot)
}
