% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StageRecordSegment.R
\name{ModelAndSegmentation.recession.regression}
\alias{ModelAndSegmentation.recession.regression}
\title{Model and segmentation of recession}
\usage{
ModelAndSegmentation.recession.regression(
  time_rec,
  daterec,
  hrec,
  uHrec,
  indx,
  nSmax = 3,
  nMin = 1,
  nCycles = 100,
  burn = 0.5,
  nSlim = max(nCycles/10, 1),
  temp.folder = file.path(tempdir(), "BaM"),
  funk = fitRecession_M3,
  ...
)
}
\arguments{
\item{time_rec}{real vector, recession duration relative to the first data detected during the recession}

\item{daterec}{vector, time in POSIXct format}

\item{hrec}{real vector, stage value of the recessions}

\item{uHrec}{real vector, uncertainty of stage}

\item{indx}{integer, factor used to gather the data of a same recession}

\item{nSmax}{integer, maximum number of segments to assess}

\item{nMin}{integer, minimum number of observations by segment}

\item{nCycles}{integer, number of MCMC adaptation cycles. Total number of simulations equal to 100*nCycles}

\item{burn}{real between 0 (included) and 1 (excluded), MCMC burning factor}

\item{nSlim}{integer, MCMC slim step}

\item{temp.folder}{directory, temporary directory to write computations}

\item{funk}{the function for estimating the recession}

\item{...}{optional arguments to funk}
}
\value{
List with the following components :
\enumerate{
\item summary_results : list with this sub components :
\itemize{

\if{html}{\out{<div class="sourceCode">}}\preformatted{ \\item summary: list, summarize the information to present to the user
 \\itemize\{
     \\item data: data frame, all data of (H,Q and uQ) with their respective periods after segmentation
     \\item shift: data frame, all detected shift time
 \}
 \\item plot : list, data formatted to use as input for some plot functions
 \\item res: list, provide all the information of the periods from tree structure
 \\itemize\{
     \\item tau: real vector, estimated shift times
     \\item segments: list, segment maximum a posterior (MAP) value indexed by the list number
     \\item mcmc: data frame, MCMC simulation
     \\item data.p: list, separate and assign information by identified stable period indexed by the list number
     \\item DIC: real, DIC estimation
     \\item nS: integer, optimal number of segments following DIC criterion
   \}
 \\item tree: data frame, provide tree structure
 \\item origin.date: positive real or date, date describing origin of the segmentation for a sample. Useful for recursive segmentation.
}\if{html}{\out{</div>}}

}
\item residuals_all_data : list, provide all information of recession data observed and simulated
\itemize{
\item time_rec: real value, time of the recession relative to first data of each recession
\item hrec: real value, stage observed
\item H_rec_sim: real value, stage simulated
\item H_res: real value, residual between stage observed and simulated
\item uhrec: real value, uncertainty on stage observed
\item uH_sim: real value, uncertainty on stage simulated
}
\item parameters : all information about the parameters estimated depending of the fit model used
}
}
\description{
Modelling recession using a catalog of fit models available to apply segmentation procedure.
Please see details for understaing how to use this function
}
\details{
To get the catalog of models available using \code{GetCatalog()}, all functions specifying Recession in the fit are concerned.
For more information about recession model and their default values, please go to the fit function, e.g. \code{?fitRecession_M3}
If prior knowledge about recession parameters is given, be sure to give dataset too.
The direction file path for this dataset must be \code{file.path(tempdir(),'BaM','Recession')} as shown in the example
Please respect the name of the parameters if default values will not be used. e.g. beta paremeter's will be "beta"
}
\examples{
recessions_extracted=Extraction_recession(H=ArdecheRiverMeyrasStage$H,
                                          uH=0.5,
                                          time=ArdecheRiverMeyrasStage$Date,
                                          chi=3,
                                          tgood=30)

recessions = recessions_extracted$Rec_extracted

# Plot all recession extracted
plot_rec_extracted(Rec_extracted = recessions)

D=RBaM::dataset(X=recessions['time_rec'],
                Y=recessions['hrec'],
                Yu=recessions['uHrec'],
                VAR.indx=recessions['indx'],
                data.dir=file.path(tempdir(),'BaM','Recession'))

# Give prior knowledge about recession-specific parameters and keep prior on stable parameters
alpha1.object = RBaM::parameter_VAR(name='alpha1',
                                    index='indx',
                                    d=D,
                                    init=rep(100,max(recessions$indx)),
                                    prior.dist=rep('Uniform',max(recessions$indx)),
                                    prior.par=rep(list(c(0,1000)),max(recessions$indx)))

alpha2.object = RBaM::parameter_VAR(name='alpha2',
                                    index='indx',
                                    d=D,
                                    init=rep(50,max(recessions$indx)),
                                    prior.dist=rep('Uniform',max(recessions$indx)),
                                    prior.par=rep(list(c(0,100)),max(recessions$indx)))

beta.object = RBaM::parameter_VAR(name='beta',
                                    index='indx',
                                    d=D,
                                    init=rep(52.4,max(recessions$indx)),
                                    prior.dist=rep('Uniform',max(recessions$indx)),
                                    prior.par=rep(list(c(-97.6,102.4)),max(recessions$indx)))

model_rec=ModelAndSegmentation.recession.regression(time_rec=recessions$time_rec,
                                                    daterec=recessions$date,
                                                    hrec=recessions$hrec,
                                                    uHrec=recessions$uHrec,
                                                    indx=recessions$indx,
                                                    alpha1.object=alpha1.object,
                                                    alpha2.object=alpha2.object,
                                                    beta.object=beta.object,
                                                    Dataset.object=D,
                                                    nCyclesrec=100,
                                                    burnrec=0.5,
                                                    nSlim=10)

# Plot recession segmentation
plot_modelAndSegm_recession(model_rec=model_rec,
                            spec_recession=c(2,16,28))

plotSegmentation(summary = model_rec$summary_results$summary,
                 plot_summary = model_rec$summary_results$plot )
}
