% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilitiesPrivate.R
\name{NumericFormatTransform}
\alias{NumericFormatTransform}
\title{Transform numeric format}
\usage{
NumericFormatTransform(numeric.date, class, origin.date)
}
\arguments{
\item{numeric.date}{real vector, numeric date}

\item{class}{string value, class of a vector (character, POSIXct, Date)}

\item{origin.date}{value, origin date in character, POSIXct or Date format}
}
\value{
numeric date to input format of time
}
\description{
Numeric date to input format of time
}
