% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotModelAndSegmentation}
\alias{plotModelAndSegmentation}
\title{Plot segmentation with the updated rating curve}
\usage{
plotModelAndSegmentation(summary)
}
\arguments{
\item{summary}{list, summary data resulting from model and segmentation function}
}
\value{
List with the following components :
\enumerate{
\item plotRC: ggplot, rating curve after segmentation
\item plotRCLog= ggplot, rating curve in log scale after segmentation
\item PlotStageSegmentation: ggplot, stage record and shift times
\item plotresidual: ggplot, residual segmentation
}
}
\description{
Plot the final segmentation with the updated rating curve,
displaying shift time along with uncertainties
}
\examples{
# Apply recursive model and segmentation function
results=recursive.ModelAndSegmentation(H=RhoneRiver$H,
                                       Q=RhoneRiver$Q,
                                       time=RhoneRiver$Year,
                                       uQ=RhoneRiver$uH)

# plot recursive model and segmentation function
plots=plotModelAndSegmentation(summary=results$summary)

# Rating curve
plots$plotRC

# Rating curve in log scale
plots$plotRCLog

# Plot stage record and shift estimated
plots$PlotStageSegmentation

# Plot final residual
plots$plotresidual
}
