% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotRCPrediction}
\alias{plotRCPrediction}
\title{Plot rating curve from a tree structure}
\usage{
plotRCPrediction(
  Hgrid = data.frame(grid = seq(-1, 2, by = 0.01)),
  autoscale = FALSE,
  temp.folder = file.path(tempdir(), "BaM"),
  CalibrationData = "CalibrationData.txt",
  allnodes = FALSE,
  nodes = 1
)
}
\arguments{
\item{Hgrid}{data frame, grid user-defined for plotting rating curve}

\item{autoscale}{logical, auto scale following data for plotting Hgrid}

\item{temp.folder}{directory, temporary directory to write computations}

\item{CalibrationData}{character, name of the calibration data used in the \code{recursive.ModelAndSegmentation} function. It must to match or an error message will be appear}

\item{allnodes}{logical, if TRUE all node will be plotted and the parameter nodes in the input data function will not be considered. If FALSE, it must to specify the nodes to plot.}

\item{nodes}{integer vector, the nodes from tree structure for plotting rating curve.}
}
\value{
ggplot, rating curve with MAP, uncertainties and gauging data. More specification in 'Details'
}
\description{
Plot all rating curve specified in the nodes after using \code{recursive.ModelAndSegmentation} function.
}
\details{
The function allows plotting any node from the tree structure, as shown in the example from \code{recursive.ModelAndSegmentation}.
The rating curves consist of the optimal rating curve after estimation, along with the parametric and total uncertainty.
Gauging data used for calibration during segmentation have been plotted. Hence, the smaller the number of the node, the more gauging data have been used for calibration.
}
