% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotResidual_ModelAndSegmentation}
\alias{plotResidual_ModelAndSegmentation}
\title{Plot residuals with updating rating curve}
\usage{
plotResidual_ModelAndSegmentation(summary)
}
\arguments{
\item{summary}{list, summary data resulting from model and segmentation function}
}
\value{
ggplot, residual segmentation
}
\description{
Plot residuals after using \code{recursive.ModelAndSegmentation} function, along with associated uncertainties
}
