% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotResidual_ModelAndSegmentation}
\alias{plotResidual_ModelAndSegmentation}
\title{Plot residuals with updating rating curve}
\usage{
plotResidual_ModelAndSegmentation(summary, plot_summary, ...)
}
\arguments{
\item{summary}{list, summary data resulting from model and segmentation function}

\item{plot_summary}{list, plot data resulting from any segmentation function}

\item{...}{optional arguments, as vector, to consider shift declared and stored by the hydrometric unit}
}
\value{
ggplot, residual segmentation
}
\description{
Plot residuals after using \code{recursive.ModelAndSegmentation} function, along with associated uncertainties
}
