% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotSegmentation}
\alias{plotSegmentation}
\title{Plot segmentation}
\usage{
plotSegmentation(data, shift)
}
\arguments{
\item{data}{list, summary data resulting from any segmentation function}

\item{shift}{list, summary shift resulting from any segmentation function}
}
\value{
a ggplot
}
\description{
Plot the final segmentation displaying shift time along with uncertainties
}
\examples{
# Apply recursive segmentation
results=recursive.segmentation(obs=RhoneRiver$H,time=RhoneRiver$Year,u=RhoneRiver$uH)

# plot recursive segmentation
plotSegmentation(data=results$summary$data,
                 shift=results$summary$shift)
}
