% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotSegmentation}
\alias{plotSegmentation}
\title{Plot segmentation}
\usage{
plotSegmentation(summary)
}
\arguments{
\item{summary}{list, summary data resulting from any segmentation function}
}
\value{
a ggplot
}
\description{
Plot the segmentation displaying shift time
}
