% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotSegmentation}
\alias{plotSegmentation}
\title{Plots segmentation}
\usage{
plotSegmentation(
  summary,
  plot_summary,
  shift_data_reported = NULL,
  show_unc_interval = FALSE
)
}
\arguments{
\item{summary}{data frame, summary data resulting from any segmentation function without estimation of rating curve}

\item{plot_summary}{list, plot data resulting from any segmentation function}

\item{shift_data_reported}{vector, shift declared and stored by the hydrometric unit}

\item{show_unc_interval}{logical, plot of posterior shift estimation as : if \code{TRUE} uncertainty interval at 95\% and if \code{FALSE} density distribution}
}
\value{
List with the following components :
}
\description{
Plots  focusing on the observed data and shift time and their estimation
}
