% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotSegmentation}
\alias{PlotSegmentation}
\title{Plots segmentation}
\usage{
PlotSegmentation(
  summary,
  plot_summary,
  shift_data_reported = NULL,
  show_unc_interval = FALSE
)
}
\arguments{
\item{summary}{data frame, summary data resulting from any segmentation function without estimation of rating curve}

\item{plot_summary}{list, plot data resulting from any segmentation function}

\item{shift_data_reported}{vector, shift declared and stored by the hydrometric unit}

\item{show_unc_interval}{logical, plot of posterior shift estimation as : if \code{TRUE} uncertainty interval at 95\% and if \code{FALSE} density distribution}
}
\value{
List with the following components :
\enumerate{
\item final_plot: ggplot, observed data presenting the shift time and their estimation
\item observation_and_shift: ggplot, observed data indexed by period, and the estimated shift time is indicated vertically,
\item shift_time_density: ggplot, density plot of the shift times following MCMC results, with
vertical lines indicating the 95\% credibility interval and a red cross representing shift time assignment
}
}
\description{
Plots  focusing on the observed data and shift time and their estimation
}
