% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plotStage_ModelAndSegmentation}
\alias{plotStage_ModelAndSegmentation}
\title{Plot shift times}
\usage{
plotStage_ModelAndSegmentation(summary, uH = NA)
}
\arguments{
\item{summary}{list, summary data resulting from model and segmentation function}

\item{uH}{vector, uncertainty concerning the observed stage}
}
\value{
ggplot, stage record and shift times
}
\description{
Plot shift times after using \code{recursive.ModelAndSegmentation} function, along with associated uncertainties
}
