% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{PlotTree}
\alias{PlotTree}
\title{Plot recursive segmentation tree}
\usage{
PlotTree(tree)
}
\arguments{
\item{tree}{data frame, tree resulting from the call of recursive.segmentation and ModelAndSegmentation_Gaugings}
}
\value{
a ggplot
}
\description{
Plot the tree resulting from any recursive segmentation procedures
}
\examples{

# Apply recursive segmentation
results=recursive.segmentation(obs=CongoRiverMINAN$Q,
                               time=CongoRiverMINAN$year,
                               u=CongoRiverMINAN$uQ,
                               nSmax=2)

# plot recursion tree
PlotTree(results$tree)
}
