% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTree}
\alias{plotTree}
\title{Plot recursive segmentation tree}
\usage{
plotTree(tree)
}
\arguments{
\item{tree}{data frame, tree resulting from the call of recursive.segmentation and recursive.ModelAndSegmentation}
}
\value{
a ggplot
}
\description{
Plot the tree resulting from any recursive segmentation procedures
}
\examples{

# Apply recursive segmentation
results=recursive.segmentation(obs=CongoRiverBrazzavilleMINAN$Q,
                               time=CongoRiverBrazzavilleMINAN$year,
                               u=CongoRiverBrazzavilleMINAN$uQ,
                               nSmax=2)

# plot recursion tree
plotTree(results$tree)
}
