% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecessionCurveEquationsModels.R
\name{Recession_B2_Equation}
\alias{Recession_B2_Equation}
\title{Recession with one exponential and four parameters}
\usage{
Recession_B2_Equation(t, alpha_1, lambda_1, c, beta_k)
}
\arguments{
\item{t}{real value, time}

\item{alpha_1}{real value, initial stage of the exponential function}

\item{lambda_1}{real value, first parameter of the exponential function}

\item{c}{real value, recession exponent}

\item{beta_k}{real value, asymptotic stage}
}
\description{
Recession with one exponential and four parameters
}
