% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecessionCurveEquationsModels.R
\name{Recession_B2_Equation}
\alias{Recession_B2_Equation}
\title{Model 2:}
\usage{
Recession_B2_Equation()
}
\value{
Parameter's description:
\enumerate{
\item{t real value, time}
\item{lambda_1 real value, first parameter of the exponential function}
\item{beta_k real value, asymptotic stage}
\item{alpha_1 real value, initial stage of the exponential function}
\item{c real value, recession exponent}
}
}
\description{
Recession with one exponential and four parameters: \deqn{\alpha \cdot \exp(-\lambda \cdot t^c) + \beta_k}
}
