% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{recursive.segmentation}
\alias{recursive.segmentation}
\title{Recursive segmentation}
\usage{
recursive.segmentation(
  obs,
  time = 1:length(obs),
  u = 0 * obs,
  nSmax = 2,
  nMin = 1,
  nCycles = 100,
  burn = 0.5,
  nSlim = max(nCycles/10, 1),
  temp.folder = file.path(tempdir(), "BaM")
)
}
\arguments{
\item{obs}{real vector, observations}

\item{time}{real vector, time}

\item{u}{real vector, uncertainty in observations (as a standard deviation)}

\item{nSmax}{integer, maximum number of segments to assess}

\item{nMin}{integer, minimum number of observations by segment}

\item{nCycles}{integer, number of MCMC adaptation cycles. Total number of simulations equal to 100*nCycles.}

\item{burn}{real between 0 (included) and 1 (excluded), MCMC burning factor}

\item{nSlim}{integer, MCMC slim step}

\item{temp.folder}{directory, temporary directory to write computations}
}
\value{
List with the following components :
\enumerate{
  \item tau: real vector, estimated shift times
  \item segments: list, segment mean value indexed by the list number
  \item mcmc: data frame, MCMC simulation
  \item data.p: list, separate and assign information by identified stable period indexed by the list number
  \item DIC: real, DIC estimation
  \item nS: integer, optimal number of segments following DIC criterion
  \item tree : data frame, table for tree structure after segmentation
}
}
\description{
Recursive segmentation procedure for a \strong{unknown} given number of segments
}
