% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{segmentation}
\alias{segmentation}
\title{Segmentation}
\usage{
segmentation(
  obs,
  time = 1:length(obs),
  u = 0 * obs,
  nSmax = 2,
  nMin = 1,
  nCycles = 100,
  burn = 0.5,
  nSlim = max(nCycles/10, 1),
  temp.folder = file.path(tempdir(), "BaM")
)
}
\arguments{
\item{obs}{real vector, observations}

\item{time}{real vector, time in POSIXct, string or numeric format}

\item{u}{real vector, uncertainty in observations (as a standard deviation)}

\item{nSmax}{integer, maximum number of segments to assess}

\item{nMin}{integer, minimum number of observations by segment}

\item{nCycles}{integer, number of MCMC adaptation cycles. Total number of simulations equal to 100*nCycles}

\item{burn}{real between 0 (included) and 1 (excluded), MCMC burning factor}

\item{nSlim}{integer, MCMC slim step}

\item{temp.folder}{directory, temporary directory to write computations}
}
\value{
list with the following components:
\enumerate{
\item summary: list, summarize the information to present to the user
\itemize{
\item data: data frame, all data with their respective periods after segmentation
\item shift: data frame, all detected shift time in numeric or POSIXct format in UTC
}
\item plot : list, data formatted to use as input for some plot functions
\item res: list, provide all the information of the periods from tree structure
\itemize{
\item tau: real vector, estimated shift times in numeric or POSIXct format in UTC
\item segments: list, segment maximum a posterior (MAP) value indexed by the list number
\item mcmc: data frame, MCMC simulation
\item data.p: list, separate and assign information by identified stable period indexed by the list number
\item DIC: real, DIC estimation
\item origin.date: positive real or date, date describing origin of the segmentation for a sample. Useful for recursive segmentation.
}
\item nS: integer, optimal number of segments following DIC criterion
}
}
\description{
Segmentation procedure for an \strong{unknown} given number of segments
}
\examples{
# Run segmentation engine function at two segments
res=segmentation(obs=RhoneRiver$H,time=RhoneRiver$Year,u=RhoneRiver$uH,nSmax=3)

# Get lower DIC value and optimal number of segments (to define optimal solution)
DIC.df = data.frame(nS=c(1:3),DIC=c(res$results[[1]]$DIC,res$results[[2]]$DIC,res$results[[3]]$DIC))
nSopt=res$nS

ggplot2::ggplot(DIC.df,ggplot2::aes(x=nS,y=DIC,col=factor(nS)))+
  ggplot2::geom_point(size=3,show.legend = FALSE)+
  ggplot2::geom_segment(ggplot2::aes(x=nSopt,y=min(DIC)*1.03,xend=nSopt,yend=min(DIC)*1.005),
                        arrow=ggplot2::arrow(length=ggplot2::unit(0.5,'cm')),
                        color='BLACK',lwd=1, show.legend = FALSE)+
  ggplot2::theme_bw()
# Data information
knitr::kable(head(res$results[[nSopt]]$summary$data),
             align = 'c',row.names = FALSE)
# Shift information
knitr::kable(head(res$results[[nSopt]]$summary$shift),
             align = 'c',row.names = FALSE)
# Plot segmentation
plotSegmentation(res$summary,
                 res$plot)
}
