% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Segmentation.R
\name{segmentation_engine}
\alias{segmentation_engine}
\title{Segmentation engine}
\usage{
segmentation_engine(
  obs,
  time = 1:length(obs),
  u = 0 * obs,
  nS = 2,
  nMin = 1,
  nCycles = 100,
  burn = 0.5,
  nSlim = max(nCycles/10, 1),
  temp.folder = file.path(tempdir(), "BaM")
)
}
\arguments{
\item{obs}{real vector, observations}

\item{time}{real vector, time}

\item{u}{real vector, uncertainty in observations (as a standard deviation)}

\item{nS}{integer, number of segments}

\item{nMin}{integer, minimum number of observations by segment}

\item{nCycles}{integer, number of MCMC adaptation cycles. Total number of simulations equal to 100*nCycles.}

\item{burn}{real between 0 (included) and 1 (excluded), MCMC burning factor}

\item{nSlim}{integer, MCMC slim step}

\item{temp.folder}{directory, temporary directory to write computations}
}
\value{
list with the following components :
\enumerate{
  \item tau: real vector, estimated shift times
  \item segments: list, segment mean value indexed by the list number
  \item mcmc: data frame, MCMC simulation
  \item data.p: list, separate and assign information by identified stable period indexed by the list number
  \item DIC: real, DIC estimation
}
}
\description{
Segmentation procedure for a \strong{known} given number of segments
}
\examples{
# Create observation vector
obs=c(rnorm(25,mean=0,sd=1),rnorm(25,mean=2,sd=1))

# Run segmentation engine function
res <- segmentation_engine(obs=obs)

# Estimated shift time
res$tau

# mean value per segment indexed by the list number
res$segments

# Uncertainty in shift time
hist(res$mcmc$tau1)

# Separate and assign information by identified stable period
res$data.p

# DIC estimation
res$DIC

# Plot
plot(obs)
lines(x=res$data.p$time.p[[1]],y=res$segments[[1]],col='blue')
lines(x=res$data.p$time.p[[2]],y=res$segments[[2]],col='blue')
abline(v=res$tau,col='green')
abline(v=quantile(res$mcmc$tau1,probs=c(0.025,0.975)),col='green',lty=2)
}
